move io functions to io_utils
